Refactored to separate files
